requirements.txt
adafruit-blinka>=8.0.0
adafruit-circuitpython-ssd1306>=2.0.0
Pillow>=10.0.0
RPi.GPIO>=0.7.0
smbus2>=0.4.0
numpy>=1.24.0